+++ drift/test/test_utils/test_utils.dart
@@ -29,3 +29,29 @@ class _NullExecutor extends Fake implements QueryExecutor {
   @override
   SqlDialect get dialect => SqlDialect.sqlite;
 }
+
+class CustomTable extends Table with TableInfo<CustomTable, Null> {
+  @override
+  final String actualTableName;
+  @override
+  final DatabaseConnectionUser attachedDatabase;
+  final List<GeneratedColumn<Object>> columns;
+  final String? _alias;
+
+  CustomTable(this.actualTableName, this.attachedDatabase, this.columns,
+      [this._alias]);
+
+  @override
+  List<GeneratedColumn<Object>> get $columns => columns;
+
+  @override
+  String get aliasedName => _alias ?? actualTableName;
+
+  @override
+  CustomTable createAlias(String alias) {
+    return CustomTable(actualTableName, attachedDatabase, columns, alias);
+  }
+
+  @override
+  Null map(Map<String, dynamic> data, {String? tablePrefix}) => null;
+}