+++ drift_dev/lib/src/writer/tables/data_class_writer.dart
@@ -53,6 +53,10 @@ class DataClassWriter {
     }
 
     // write constructor with named optional fields
+
+    if (!scope.options.generateMutableClasses) {
+      _buffer.write('const ');
+    }
     _buffer
       ..write(table.dartTypeName)
       ..write('({')